+++ app/views/index.blade.php
@@ -123,8 +123,14 @@
                                     </div>
 
                                 {{ Form::close() }}
+                                <script>
 
-                                {{ Form::open(array('action' => array('NetworkController@deleteNetwork', $network->id), 'class' => 'form-horizontal', 'method'=>'DELETE')) }}
+                                  function ConfirmDelete(){
+                                      return confirm("Are you sure you want to delete the network {{ $network->name }}?");
+                                  }
+
+                                </script>
+                                {{ Form::open(array('action' => array('NetworkController@deleteNetwork', $network->id), 'class' => 'form-horizontal', 'method'=>'DELETE', 'onsubmit' => 'return ConfirmDelete()')) }}
                                     <div style="margin-top:10px" class="form-group">
                                         <div class="col-md-12">
                                             {{ Form::submit('Delete', array('class'=>'btn btn-danger')) }}
